refactor to remove layout_utils.py
